Bugfixes in deserializing controllers
git-svn-id: http://elektro.robotics.tu-berlin.de/svn/rswin/trunk@872 915fd5a2-32c1-4e06-90b4-f1382ce795f9

+++ src/hybrid_automaton/HybridAutomaton.cpp
@@ -59,7 +59,7 @@ MotionBehaviour* HybridAutomaton::getNextMotionBehaviour(const Milestone* curren
 				{
 					if(bad_edges)
 					{
-						for(int j = 0; j < bad_edges->size(); j++)
+						for(size_t j = 0; j < bad_edges->size(); j++)
 						{
 							// problem: very often the edges are from current_ms!!!
 							if((*bad_edges)[j].first == currentMs->getName() && (*bad_edges)[j].second == ((Milestone*)edges[i]->getChild())->getName())
